24/7/7: Add base support for EEPROM

+++ iFOC-Port/STM32/hw_i2c_interface.hpp
@@ -9,8 +9,8 @@ class HWI2C : public I2CBase
 public:
     HWI2C(I2C_HandleTypeDef *_hi2c) : hi2c(_hi2c) {};
     void Init() override;
-    uint8_t WriteBytes(uint8_t addr, uint8_t *data, uint8_t len) override; // return 1 if success, else 0
-    uint8_t ReadBytes(uint8_t addr, uint8_t *data, uint8_t len) override;
+    uint8_t WriteBytes(uint8_t addr, uint8_t *data, uint16_t len) override; // return 1 if success, else 0
+    uint8_t ReadBytes(uint8_t addr, uint8_t *data, uint16_t len) override;
     void DelayMs(uint32_t ms) override { delay_ms(ms); };
 private:
     I2C_HandleTypeDef *hi2c;
@@ -21,7 +21,7 @@ void HWI2C::Init()
     
 }
 
-uint8_t HWI2C::WriteBytes(uint8_t addr, uint8_t *data, uint8_t len)
+uint8_t HWI2C::WriteBytes(uint8_t addr, uint8_t *data, uint16_t len)
 {
     // addr &= ~0x01; // R/W bit
     // __disable_irq();
@@ -30,7 +30,7 @@ uint8_t HWI2C::WriteBytes(uint8_t addr, uint8_t *data, uint8_t len)
     return ret == HAL_OK;
 }
 
-uint8_t HWI2C::ReadBytes(uint8_t addr, uint8_t *data, uint8_t len)
+uint8_t HWI2C::ReadBytes(uint8_t addr, uint8_t *data, uint16_t len)
 {
     // __disable_irq();
     HAL_StatusTypeDef ret = HAL_I2C_Master_Receive(hi2c, (uint16_t)(addr << 1), data, len, 0xFFF);
